Move hyper legacy client i13n into its own mod
